Add files via upload
bank_balanced.csv contains 10578 records of a Portuguese bank. The data is related with direct marketing campaigns. The classification goal is to predict if the client will subscribe a term deposit (variable y).

Problem Steps:

1. Read data and select the fields that you think can help with better prediction. (10 pt)

Note: selection of all fields is certainly an option).

2. Change categorical values to numeric and scale the values. (30 pt)

Note: you can do this using skitlearn OrdinalEncoder and StandardScaler.

3. Divide the data into training set and test sets. (10 pt)

4. Choose any of machine learning algorithms dicussed in class (Decision Tree, Random Forest, GBM) to train a predictive model using training set. (30)

5. Evaluate the model with accuracy and confusion matrix. (20 pt)

Bounce: Find a model with high accuracy over 90% with very low false positive and false negatives. (50 pt)

 

About Data:

This is a balanced dataset based on a Kaggle's binary classification problem (Links to an external site.). The dataset does not have any missing values. Some of the fields are also removed for simplicity. Here is the data dictionary for this dataset:

1 - age (numeric)
2 - job : type of job (categorical: "admin.","unknown","unemployed","management","housemaid","entrepreneur","student",
"blue-collar","self-employed","retired","technician","services")
3 - marital : marital status (categorical: "married","divorced","single"; note: "divorced" means divorced or widowed)
4 - education (categorical: "unknown","secondary","primary","tertiary")
5 - default: has credit in default? (binary: "yes","no")
6 - balance: average yearly balance, in euros (numeric)
7 - housing: has housing loan? (binary: "yes","no")
8 - loan: has personal loan? (binary: "yes","no")

9 - contact: contact communication type (categorical: "unknown","telephone","cellular")
10 - day: last contact day of the month (numeric)
11 - month: last contact month of year (categorical: "jan", "feb", "mar", …, "nov", "dec")
12 - duration: last contact duration, in seconds (numeric)